Add files via upload
I Developed a dashboard that displays KPIs for informed decision-making. Sam Bikeshare wants to understand if they can increase subscription prices for the next year.

Dashboard Requirements
Hourly revenue analysis
Profit and revenue trends
Seasonal revenue (quarters)
Rider demographics


Conclusion
Despite the price increase, there was an increase in the number of registered and casual riders.
Promotions can be offered during the third-quarter months as that is when usage peaked. This can lead to an increase in profits and the profit margin
The market can be tested with a 5–10% price increase.